configure: [Windows] Add --with-winver option to specify minimum Windows version
This effectively defines WINVER and _WIN32_WINNT to the specified value
which will compile the code with that minimum supported Windows version
instead of the compiler/SDK default.
The passed value is expected to be in hex (like 0x0600) or decimal, so
passing --with-winver=0x0600 is a valid accepted input. See
https://learn.microsoft.com/cpp/porting/modifying-winver-and-win32-winnt
for the values of different Windows versions.