utils: Correct ctype(3) usage
According to POSIX.1 the argument of ctype(3) functions, including
toupper(3), has to either be EOF or a value representable as an unsigned
char. If the argument has any other value, the behavior is undefined. This
means values of signed char must first be cast to unsigned char, otherwise
they will be outside the range of allowed values.

glibc attempts to avoid the undefined behavior by defining the functions to
work for all integer inputs representable by either an unsigned char or a
signed char. NetBSD libc does not, and causes a segfault if the value is
invalid.